Modify dke_preprocess_prisma to use basedir as the location of DICOM data
The original script from which dke_preprocess_prisma came looped over
subject directories, which were hard-coded. Each subject directory was
assumed to have a subdirectory called 'dicom_dke' which had the DICOM
data.

In DKE, the study directory and the subject subdirectories are specified
in the parameter file, and looping over the subjects is done in dke.m.
If the input image format is DICOM, dke_preprocess_dicom is called
with the path of the current subject directory as the first argument,
basedir. dke_preprocess_dicom creates a subdirectory called
'intermediate_processing', and the DICOM images are converted to
NIfTI format and preprocessed within that subdirectory.

Now dke_preprocess_prisma also uses the argument basedir as the
location for DICOM images. It creates a subdirectory called
'intermediate_processing', and the DICOM images are converted to
NIfTI format and preprocessed within that subdirectory.

After running dke_preprocess_prisma, the directory structure is
        basedir (location of DICOM images = first argument of function)
          \--> intermediate_processing
                 |--> nifti
                 |       |--> B0
                 |       |--> DKI1
                 |       \--> combined
                 \--> dke