Get b values from DICOM headers
The code for reading the b values comes from the function nii_name
in spm_dicom_convert_series_description.m.

Previously in dke_preprocess_prisma.m the b values were hard-coded:
64 directions, b=1000 and 2000, and 9 extra b=0 images = parameters
from the data set used for testing.

This commit includes a test that the b values read from the DICOM
headers are the same as those that were hard-coded.