Get name of the file to preprocess from series_description parameter
Previously dke_preprocess_prisma took the first file that matched
'*DKI*.nii' as the file to preprocess. Now it gets the name from
the first element of preprocess_options.series_description in the
parameter file, with '.nii' appended.

Also the dcm2niix command was changed to use the series description
(rather than the protocol name) plus '.nii' for the output file
name.