Remove hard-coding of initial name of current_4D_file
Now the initial name of the image file being preprocessed is the
same as that given by the first element of the series description
parameter in the parameter file, with '.nii' appended. This is the
same as the name of the NIfTI file output from dcm2niix.

Also renamed bval_file1 to bval_file.