Modify b0_dir variable to be associated with the DKI series
The code currently handles one DKI series, but it will soon be
modified to handle more. The b0_dir directory is for b=0 files
from a DKI series, namely the first volume and all other volumes
which have b=0. Since there could be more than one DKI series,
there needs to be a separate b0_dir directory for each series.

The current code averages the files in the b0_dir directory,
but not the files in the extra_b0 directory, and it doesn't
coregister the files.